Localization for pretty list items
This commit adds support to forms for the model, its attributes and
refactors flash messages.

Got to stay away from using the reserved "models" key. This caused some
problems with rendering the submit button via the form helper when the
models key was present along with the destroy key that accepts a name
argument.